feat: make request and configure navigation controller
